refactor : api 분리 (#5)
