+++ src/pages/Login.tsx
@@ -3,7 +3,7 @@ import { useMutation } from "react-query";
 import { useForm } from "react-hook-form";
 import styled, { css } from "styled-components";
 
-import { login } from "apis/axios";
+import { login } from "apis";
 
 const Login = () => {
   const navigate = useNavigate();
